requirements.txt
kafka-python==2.0.2
av==11.0.0